[CARMEL-2827] Fix NumberFormatException for delta metrics (#57)

+++ src/main/scala/org/apache/spark/sql/delta/DeltaOperations.scala
@@ -187,7 +187,7 @@ object DeltaOperations {
         strMetrics += "numRowsCopied" -> "0"
       } else {
         strMetrics += "numRowsCopied" -> (
-          numOutputRows - strMetrics("numRowsUpdated").toInt).toString
+          numOutputRows - strMetrics("numRowsUpdated").toLong).toString
       }
       strMetrics
     }
